window finder reorganize
