+++ satellite_determination/dataclasses/overhead_window.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+from satellite_determination.dataclasses.satellite import Satellite
+from satellite_determination.dataclasses.time_window import TimeWindow
+
+
+@dataclass
+class OverheadWindow:
+    satellite: Satellite
+    overhead_time: TimeWindow
