--- tests/window_finder/satellite_jsons/satellites_empty.json
@@ -1,3 +1,0 @@
-{
-  "satellites": []
-}
