--- tests/window_finder/satellite_jsons/satellites_some_overhead.json
@@ -1,12 +1,0 @@
-{
-  "satellites": [
-    {
-      "frequency": {
-        "high_in_megahertz": 2.0,
-        "low_in_megahertz": 1.0
-      },
-      "satellite_number": "ARBITRARY1",
-      "name": "ARBITRARY1"
-    }
-  ]
-}
